requirements.txt
# Core dependencies
strands-agents>=1.0.0
boto3>=1.34.0
requests>=2.31.0
python-dateutil>=2.8.2
pydantic>=2.5.0
gnews>=0.3.7
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# AWS Lambda dependencies
aws-lambda-powertools>=2.25.0